// how-to/here-supertabs-starter/client/src/framework/shapes/interopbroker-shapes.ts
export interface PlatformInteropOverride<O = unknown, H = ModuleHelpers> extends ModuleImplementation<O, H> {
	/**
	 * Get the override constructor for the interop broker (useful if you wish this implementation to be layered with other implementations and passed to the platform's initialization object as part of an array).
	 * @param options The options for the interop broker defined as part of the platform.
	 * @returns The override constructor to be used in an array.
	 */
	getConstructorOverride(
		options: PlatformInteropOverrideOptions
	): Promise<OpenFin.ConstructorOverride<OpenFin.InteropBroker>>;
}

/**
 * The options passed to the platform interop broker override.
 */
export type PlatformInteropOverrideOptions = Omit<PlatformInteropBrokerOptions, "modules"> & {
	/**
	 * The platform's window positioning options that can be passed to the getWindowPositionUsingStrategy function along with an identity to determine where a window should be placed.
	 */
	windowPositionOptions?: WindowPositioningOptions;
	/**
	 * The platform's root url that can be used to resolve relative urls.
	 */
	platformRootUrl?: string;
};

/**
 * Options for the platform interop broker.
 */
export interface PlatformInteropBrokerOptions extends ModuleList {
	/**
	 * Intent Resolver configuration if you wish to support intents. It needs to support the functions required by the
	 * platform
	 */
	intentResolver?: IntentResolverOptions;

	/**
	 * Options related to the way this platform supports intents
	 */
	intentOptions?: IntentOptions;

	/**
	 * When fdc3.open is used what settings should be applied?
	 */
	openOptions?: OpenOptions;

	/**
	 * Options for when fdc3.broadcast or fin.me.interop.setContext is called.
	 */
	contextOptions?: ContextOptions;

	/**
	 * When fdc3.getInfo is used what settings should be taken into account?
	 */
	getInfoOptions?: GetInfoOptions;

	/**
	 * If an unregistered app is included here then it indicates you wish to support selecting views/windows that are
	 * not linked to an app from an intent picker that supports instances. The intents and contexts in this app specify
	 * which you support for unregistered instances. Do not specify a manifest or manifestType for this entry (we don't
	 * launch unregistered apps so there is no need for a manifest as it is existing instances and we specify a custom
	 * manifestType for consideration when we are bringing an unregistered web app instance into focus).
	 */
	unregisteredApp?: PlatformApp;
}
// ...
export interface IntentOptions {
	/**
	 * How long should the broker wait after launching a view/window for it to register an intent handler. The default
	 * is 15000 (15 seconds)
	 */
	intentTimeout?: number;
}

/**
 * Option for the Context Handling.
 */
export interface ContextOptions {
	/**
	 * Should the broker send context messages back to the sender? Default is true.
	 */
	includeOriginator?: boolean;
}

/**
 * Get Info Options.
 */
export interface GetInfoOptions {
	/**
	 * Should the broker include the interop capabilities as specified in the app directory within the appMetadata so that the app can take into account what the platform has specified should be supported (and whether there is a misconfiguration where the platform's app definition doesn't reflect the capabilities of the app).
	 * Default is false as this is not part of the fdc3 standard. It is placed within appMetadata.instanceMetadata.interop rather than appending it to the root of appMetadata as it is not part of the fdc3 standard.
	 */
	includeAppInteropInfo?: boolean;

	/**
	 * Should the broker include the full app meta data or just appId and instanceId (and interopInfo if enabled). Default is false.
	 */
	includeAllAppMetadataInfo?: boolean;
}

/**
 * Intent resolver options.
 */
export interface IntentResolverOptions {
	/**
	 * The url of the html page that has the intent picker
	 */
	url: string;

	/**
	 * the height you wish the window to be
	 */
	height?: number;

	/**
	 * the width you wish the window to be
	 */
	width?: number;

	/**
	 * the fdc3 api version this picker will support (default is v2)
	 */
	fdc3InteropApi?: string;

	/**
	 * A suggested title for the intent picker/resolver ui
	 */
	title?: string;
}
// ...
export interface OpenOptions {
	/**
	 * When fdc3.open is raised will it only apply to applications that support the intent "OpenApp" (context if passed is sent to the intent handler for OpenApp)
	 * or will it follow the fdc3 approach where all apps can be opened and the defaultContextListener will receive the context if passed.
	 * The default is fdc3. The previous behavior was "intent" and you can set this setting in order to have fdc3 open only apply to apps that say they support the
	 * intent "OpenApp". This setting is here to let you keep the previous behavior.
	 */
	openStrategy?: "intent" | "fdc3";

	/**
	 * How long should the broker wait after launching a view/window for it to register a context handler. The default
	 * is 15000 (15 seconds)
	 */
	contextTimeout?: number;

	/**
	 * How long should the broker wait after launching a view/window for it to connect to the broker. The default
	 * is 15000 (15 seconds).
	 */
	connectionTimeout?: number;
}

/**
 * Interop Broker helpers provides environment methods and data.
 */
export type PlatformInteropBrokerHelpers = ModuleHelpers;